% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignsector.R
\name{assignsector}
\alias{assignsector}
\title{assignsector}
\usage{
assignsector(df = NULL)
}
\arguments{
\item{df}{A dataframe with a fleet code column (SFLT_DESC_ID)}
}
\value{
a data.frame with a new column for fleet/sector
}
\description{
Assign fleet sectors to data.
}
\examples{
example1 <- isdbpull(year=2023)
}
