% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignzone.R
\name{assignzone}
\alias{assignzone}
\title{assignzone}
\usage{
assignzone(
  marfis.df = NULL,
  isdb.df = NULL,
  y = year,
  lat.field = "LAT",
  lon.field = "LON",
  dir = "C:/LocalDataDump"
)
}
\arguments{
\item{marfis.df}{A dataframe of MARFIS data with latitude and longitude columns.}

\item{isdb.df}{A dataframe of ISDB data for the relevant year. Default is "isdb" as supplied by the function isdbpull.R}

\item{y}{Year}

\item{lat.field}{Latitude field, in decimal degrees. Default is "LAT".}

\item{lon.field}{Longitude field, in decimal degrees. Default in "LON".}

\item{dir}{Directory with locally stored data, which can be generated with the function localcopies.R. Default is "C:/LocalDataDump"}
}
\value{
a list of two dataframes: one with an additional column for Zone, with all Zones assigned; another with an additional column for Zone, with no Zones assigned.
}
\description{
Assign discard zones to data.
}
\examples{
example1 <- assignzone(marfis)
}
