% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coveragesummary.R
\name{coveragesummary}
\alias{coveragesummary}
\title{coveragesummary}
\usage{
coveragesummary(marfis.df = NULL, y = year)
}
\arguments{
\item{marfis.df}{A dataframe of MARFIS data with observer TRIP ID codes.}

\item{y}{Year}
}
\value{
An Excel document (Coverage_Summary_year.xlsx) of observer coverage in the working directory.
}
\description{
Export an observer coverage summary Excel document.
}
\examples{
example1 <- coveragesummary(marfis.df=marfis,y=2023)
}
