% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DO_BS.R
\name{DO_BS}
\alias{DO_BS}
\title{Bootstrapped Landings Multiplier}
\usage{
DO_BS(data = ready, file = "Discards_GroupedData.csv", nboot = 1000)
}
\arguments{
\item{data}{The output from the readydata function. Default is ready.}

\item{file}{The location of the file with the fleet/zone/quarter combinations. Default is "Discards_GroupedData.csv." This is the readydata output Discards_DataforGrouping_year.csv with an additional column for group.}

\item{nboot}{The number of bootstrap iterations required. Default is 1000.}
}
\value{
A dataframe with the bootstrapped multiplier, variance, and bias, percentile confidence distribution at 0.05, bias corrected percentile confidence distribution at 0.05, and whether or not the discards are significant for a given group (>0.05).
}
\description{
Bootstrap the landings multiplier and calculate variance and bias
}
\examples{
example1 <- DO_BS(data=ready)
}
