% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DO_RATIOS.R
\name{DO_RATIOS}
\alias{DO_RATIOS}
\title{DO_RATIOS}
\usage{
DO_RATIOS(df = NULL)
}
\arguments{
\item{df}{A dataframe of trips with numeric variables for cod landings and  haddock landings (called COD and HAD), and a character (Y/N) variable for whether the trip was observed (called OBS).}
}
\value{
a data.frame with landings multiplier
}
\description{
Calculate the landings ratio of cod:haddock on observed vs. unobserved trips according to the method from Gavaris et al. (2007).
}
\examples{
COD=runif(4)
HAD=runif(4)
OBS=c("N","N","Y","Y")
df<-data.frame(COD,HAD,OBS)
example1 <- DO_RATIOS(df)
}
