% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localcopies.R
\name{localcopies}
\alias{localcopies}
\title{localcopies}
\usage{
localcopies(
  year = NULL,
  dir = "C:/LocalDataDump",
  username = oracle.username,
  password = oracle.password,
  dsn = oracle.dsn,
  pkg = "roracle"
)
}
\arguments{
\item{year}{The year required for VMS data (all years are pulled for MARFIS and ISDB)}

\item{dir}{The local directory where the data will be stored. Default is "C:/LocalDataDump"}

\item{username}{Oracle username. Default is the value oracle.username stored in .Rprofile.}

\item{password}{Oracle password. Default is the value oracle.password stored in .Rprofile.}

\item{dsn}{Oracle dsn. Default is the value oracle.dsn stored in .Rprofile.}
}
\value{
.Rdata copies of MARFIS and ISDB databases, and .rds copies of relevant fleet and VMS data (mobile and fixed)
}
\description{
A function to pull local copies of the MARFIS and ISDB databases, as well as relevant fleet and VMS data.
}
