% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noncommercial.R
\name{noncommercial}
\alias{noncommercial}
\title{noncommercial}
\usage{
noncommercial(marfis.df = NULL, isdb.df = NULL)
}
\arguments{
\item{marfis.df}{A dataframe of MARFIS data with TRIP codes.}

\item{isdb.df}{A dataframe of ISDB data with TRIP codes and SETCD_ID. Default is "isdb" as supplied by the function isdbpull.R}
}
\value{
a list of two dataframes: a data frame of commercial trips to keep; another with non-commercial trips to remove
}
\description{
Remove trips that were not commercial.
}
\examples{
example1 <- noncommercial(marfis.df=marfis,isdb.df=NULL)
}
