% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsquarters.R
\name{obsquarters}
\alias{obsquarters}
\title{obsquarters}
\usage{
obsquarters(marfis.df = NULL)
}
\arguments{
\item{marfis.df}{A dataframe of MARFIS data with observer TRIP ID codes.}
}
\value{
a list of two data frames: data that are part of quarters with 100\% observer coverage (by fleet/sector and zone) to remove; another that are part of quarters with <100\% observer coverage (by fleet/sector and zone)
}
\description{
Identify quarters with 100\% observer coverage, and then remove
}
\examples{
example1 <- obsquarters(marfis.df=marfis)
}
