% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readydata.R
\name{readydata}
\alias{readydata}
\title{readydata}
\usage{
readydata(
  year = NULL,
  directory = getwd(),
  username = oracle.username,
  password = oracle.password,
  dsn = oracle.dsn
)
}
\arguments{
\item{year}{The year required.}

\item{directory}{Directory where output files will be saved. Default is the working directory.}

\item{username}{Oracle username. Default is the value oracle.username stored in .Rprofile.}

\item{password}{Oracle password. Default is the value oracle.password stored in .Rprofile.}

\item{dsn}{Oracle dsn. Default is the value oracle.dsn stored in .Rprofile.}
}
\value{
a data frame of clean data ready for analysis
}
\description{
A function to pull raw data, clean it, and QAQC it. After, it is ready for analysis.
}
\examples{
example1 <- isdbpull(year=2023)
}
