% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciessought.R
\name{speciessought}
\alias{speciessought}
\title{speciessought}
\usage{
speciessought(
  marfis.df = NULL,
  username = oracle.username,
  password = oracle.password,
  dsn = oracle.dsn,
  y = year
)
}
\arguments{
\item{marfis.df}{A dataframe of MARFIS data with TRIP codes.}

\item{username}{Oracle username. Default is the value oracle.username stored in .Rprofile.}

\item{password}{Oracle password. Default is the value oracle.password stored in .Rprofile.}

\item{dsn}{Oracle dsn. Default is the value oracle.dsn stored in .Rprofile.}

\item{y}{Year}
}
\value{
a list of two dataframes: one with trips that sought cod or haddock or keep; another with trips that did not seek cod or haddock to remove
}
\description{
Remove trips that did not seek cod or haddock
}
\examples{
example1 <- speciessought(marfis.df=marfis)
}
